# Add possibility to execute multiple slice queries together to KeyColumnValueStore [cql-tests] [tp-tests]