… docs/changelog.md
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/BackendTransaction.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/KeyMultiQuery.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/KeySliceQuery.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/KeysQueriesGroup.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/MultiKeyMultiQuery.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/MultiKeysQueryGroups.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/MultiQueriesByKeysGroupsContext.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/SliceQuery.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/StoreFeatures.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/cache/ExpirationKCVSCache.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/keycolumnvalue/cache/KCVSCache.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/util/CompletableFutureUtil.java
… janusgraph-core/src/main/java/org/janusgraph/diskstorage/util/MetricInstrumentedStore.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/database/StandardJanusGraph.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/query/BackendQueryHolder.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/query/profile/QueryProfiler.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/query/profile/QueryProfilerUtil.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/query/vertex/MultiVertexCentricQueryBuilder.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/query/vertex/VertexCentricQueryBuilder.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/transaction/StandardJanusGraphTx.java
… janusgraph-core/src/main/java/org/janusgraph/graphdb/util/MultiSliceQueriesGroupingUtil.java
… janusgraph-core/src/test/java/org/janusgraph/graphdb/util/MultiSliceQueriesGroupingUtilTest.java
… janusgraph-cql/src/main/java/org/janusgraph/diskstorage/cql/CQLKeyColumnValueStore.java
… janusgraph-inmemory/src/test/java/org/janusgraph/graphdb/inmemory/InMemoryGraphTest.java
… janusgraph-test/src/main/java/org/janusgraph/diskstorage/cache/KCVSCacheTest.java