@porunov 2023-06-15
**Benchmarks of `CQLMultiQueryMultiSlicesBenchmark`**

**`master` branch:**
```
Benchmark                                                                                                          (verticesAmount)  Mode  Cnt     Score     Error  Units
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                   5000  avgt    5    74.543 ±   7.149  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                  50000  avgt    5   832.826 ±  48.409  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                     5000  avgt    5    68.687 ±   4.614  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                    50000  avgt    5   739.975 ±  16.342  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection              5000  avgt    5   350.945 ±  15.501  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection             50000  avgt    5  3684.055 ± 125.631  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                                5000  avgt    5   278.037 ±   9.099  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                               50000  avgt    5  3177.082 ± 127.442  ms/op
```

**Current PR:**
```
Benchmark                                                                                                          (verticesAmount)  Mode  Cnt     Score     Error  Units
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                   5000  avgt    5    81.126 ±   5.636  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                  50000  avgt    5   813.297 ±  24.664  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                     5000  avgt    5    67.659 ±   2.313  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                    50000  avgt    5   737.676 ±  18.572  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection              5000  avgt    5   266.972 ±  13.312  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection             50000  avgt    5  2796.589 ±  60.584  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                                5000  avgt    5   269.096 ±   7.659  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                               50000  avgt    5  3135.563 ± 174.468  ms/op
```

**Conclusion:**
As expected `getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` and `getValuesAllPropertiesWithUnlimitedBatch` have the same performance both for `master` branch and in this PR.
We see that `getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` is 24% faster in this PR than in `master` branch. This is due to the fact that in this PR CQL query computes each slice query per each key in parallel but it is sequential operation in `master` branch.
Why didn't we see any performance improvement for `getValuesMultiplePropertiesWithUnlimitedBatch` then? This test specifically shows that performance is equal between `master` branch and the current PR for this case because CQL is configured to accept up to 1024 CQL queries in parallel, but in `master` branch we are sending 50000 CQL requests 10 times sequentially and in this PR we are sending 500000 CQL requests in parallel. 
Logically, if the driver can process only 1024 requests in parallel - it doesn't have any difference if we send 50000 CQL requests in 10 sequential batches or if we send 500000 CQL requests.

Another observation, even so `getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` is 24% faster than in `master` branch, this test is still not as fast as `getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` (which is about `3.4` times faster). 
This shows us that sending a single CQL request which retrieves all data is more beneficial than sending multiple CQL requests which retrieves the same data. Thus, after this PR is merged, we should work on #3816 to group all slice queries into a single slice query for the respective keys.