@porunov 2023-06-17
**Latest benchmarks.**

**`master` branch:**
```
Benchmark                                                                                                          (verticesAmount)  Mode  Cnt      Score     Error  Units
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                   5000  avgt    5     76.508 ±   6.666  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                  50000  avgt    5    844.627 ±  45.378  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                     5000  avgt    5     68.332 ±   1.970  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                    50000  avgt    5    746.349 ±  26.421  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection              5000  avgt    5    352.094 ±  22.025  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection             50000  avgt    5   3776.992 ± 202.947  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithSmallBatch                                                    5000  avgt    5   1002.613 ±  46.442  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithSmallBatch                                                   50000  avgt    5  10208.184 ± 133.210  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                                5000  avgt    5    289.104 ±  35.600  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                               50000  avgt    5   3188.636 ± 270.826  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesThreePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                 5000  avgt    5    111.704 ±   8.093  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesThreePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                50000  avgt    5   1186.497 ±  69.885  ms/op
CQLMultiQueryMultiSlicesBenchmark.vertexCentricPropertiesFetching                                                              5000  avgt    5   4291.133 ± 144.476  ms/op
CQLMultiQueryMultiSlicesBenchmark.vertexCentricPropertiesFetching                                                             50000  avgt    5  42870.648 ± 978.082  ms/op
```

**This PR:**
```
Benchmark                                                                                                          (verticesAmount)  Mode  Cnt      Score     Error  Units
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                   5000  avgt    5     76.598 ±   5.908  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                  50000  avgt    5    838.959 ±  44.656  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                     5000  avgt    5     69.015 ±   1.115  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesAllPropertiesWithUnlimitedBatch                                                    50000  avgt    5    748.797 ±  41.574  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection              5000  avgt    5    267.932 ±  16.651  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection             50000  avgt    5   2859.473 ± 111.065  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithSmallBatch                                                    5000  avgt    5    352.522 ±  12.514  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithSmallBatch                                                   50000  avgt    5   3740.955 ± 138.963  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                                5000  avgt    5    272.325 ±   9.806  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesMultiplePropertiesWithUnlimitedBatch                                               50000  avgt    5   3112.767 ± 324.889  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesThreePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                 5000  avgt    5     96.613 ±   5.981  ms/op
CQLMultiQueryMultiSlicesBenchmark.getValuesThreePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection                50000  avgt    5    993.698 ±  57.113  ms/op
CQLMultiQueryMultiSlicesBenchmark.vertexCentricPropertiesFetching                                                              5000  avgt    5   1002.535 ±  14.568  ms/op
CQLMultiQueryMultiSlicesBenchmark.vertexCentricPropertiesFetching                                                             50000  avgt    5  10329.903 ± 176.009  ms/op
```

**Conclusion:** 
For cases when CQL driver is overloaded with other parallel requests - performance stays the same as currently at `master` branch. However, performance quite improves for multi-slice scenarios when CQL driver is less busy.

Ideally, we need to aim to reach the same performance over fetching all vertices with a single slice query vs fetching all properties separately during same fully loaded CQL driver. I.e. `getValuesMultiplePropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` and `getValuesAllPropertiesWithAllMultiQuerySlicesUnderMaxRequestsPerConnection` ideally would perform the same.
However, this isn't the case as for now. In `master` branch the difference is `4.6` times between those tests. In this PR the difference is `3.5` times. This PR is just a step forward toward achieving same performance for these cases. Nevertheless, this PR brings at least 24% improvement and in less loaded scenarios the difference is even more (Like 2.85 times for `getValuesMultiplePropertiesWithSmallBatch` or 4.2 times for `vertexCentricPropertiesFetching`). 

The follow up performance improvement issue is #3816 , but that issue is less trivial.